refactor + add hidden fields
